+++ MAM/src/main/java/com/cafelivro/mam/workorder/DownLoadAsyncTask.java
@@ -1,9 +1,7 @@
 package com.cafelivro.mam.workorder;
 
-import android.app.Activity;
 import android.content.ContentValues;
 import android.content.Context;
-import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.AsyncTask;
@@ -13,19 +11,21 @@ import android.widget.ProgressBar;
 import android.widget.Toast;
 
 import com.cafelivro.mam.R;
+import com.cafelivro.mam.util.AsyncTaskConstants;
+import com.cafelivro.mam.util.CommonValidator;
 import com.cafelivro.oslc.MaximoConnector;
 import com.ibm.json.java.JSONArray;
 import com.ibm.json.java.JSONObject;
-
-import java.util.Iterator;
 
 /**
  * Created by baeks on 9/4/2016.
  */
 
- class DownLoadAsyncTask extends AsyncTask<Void,String,Void>{
+ class DownLoadAsyncTask extends AsyncTask<String,String,Integer>{
+
     private WorkorderListActivity activity;
     private ProgressBar progressBar;
+
     public DownLoadAsyncTask(Context context) {
         this.activity=(WorkorderListActivity)context;
         this.progressBar=(ProgressBar)activity.findViewById(R.id.progressBar);
@@ -39,19 +39,26 @@ import java.util.Iterator;
 
     }
 
-    @Override
-    protected Void doInBackground(Void... voids) {
 
-        SQLiteDatabase database=activity.openOrCreateDatabase(activity.getString(R.string.database_name), Context.MODE_PRIVATE,null);
+    @Override
+    protected Integer doInBackground(String... args) {
+        MaximoConnector connector = new MaximoConnector(activity);
+
+        if(!CommonValidator.wifiOn(activity)){
+            return AsyncTaskConstants.RETURN_WIFI_FAIL;
+        }
+
+        if(connector.canConnect("oslcwodetail")==null){
+            return AsyncTaskConstants.RETURN_CONNECT_FAIL;
+        }
 
         String oslcWhere   = "oslc.where=spi:siteid=\"BEDFORD\"";
         String oslcSelect  = "oslc.select=spi:wonum,spi:siteid,spi:description,spi:siteid,spi:assetnum";
 
-        MaximoConnector connector = new MaximoConnector(activity);
         JSONObject result=connector.query("oslcwodetail",oslcWhere,oslcSelect);
         JSONArray assetSet =(JSONArray)result.get("rdfs:member");
 
-
+        SQLiteDatabase database=activity.openOrCreateDatabase(activity.getString(R.string.database_name), Context.MODE_PRIVATE,null);
         database.execSQL("delete from workorder");
         ContentValues values;
         for(int i=0;i<assetSet.size();i++){
@@ -90,9 +97,7 @@ import java.util.Iterator;
         database.close();
 
 
-
-
-        return null;
+        return AsyncTaskConstants.RETURN_WIFI_FAIL;
     }
 
 
@@ -103,12 +108,18 @@ import java.util.Iterator;
     }
 
     @Override
-    protected void onPostExecute(Void result) {
+    protected void onPostExecute(Integer result) {
         super.onPostExecute(result);
         progressBar.setVisibility(View.INVISIBLE);
-        activity.adapter.setDataSet();
-        activity.adapter.notifyDataSetChanged();
-        Toast.makeText(activity,"DOWNLOAD 가 완료 되었습니다.",Toast.LENGTH_SHORT).show();
+        if(result==AsyncTaskConstants.RETURN_SUCCESS){
+            activity.adapter.setDataSet();
+            activity.adapter.notifyDataSetChanged();
+            Toast.makeText(activity,"DOWNLOAD IS SUCCESS",Toast.LENGTH_SHORT).show();
+        }else if(result==AsyncTaskConstants.RETURN_WIFI_FAIL){
+            Toast.makeText(activity,"WIFI IS NOT CONNECTED",Toast.LENGTH_SHORT).show();
+        }else if(result==AsyncTaskConstants.RETURN_CONNECT_FAIL){
+            Toast.makeText(activity,"REQUEST IS WRONG",Toast.LENGTH_SHORT).show();
+        }
 
     }
